Add error dialog to CheckLogWorker
An error dialog has been added to the CheckLogWorker class. Now, when an error occurs while adding log time, the user will be notified with a GUI message box, and not just a console error log. This enhancement was introduced to improve user feedback and alert users when an operation fails.